+++ Fabric/src/main/kotlin/gay/object/hexdebug/fabric/HexDebugClientAbstractionsImpl.kt
@@ -2,8 +2,11 @@
 
 package gay.`object`.hexdebug.fabric
 
+import gay.`object`.hexdebug.fabric.interop.FabricHexDebugHexicalInterop
 import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper
+import net.fabricmc.loader.api.FabricLoader
+import net.minecraft.client.Minecraft
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.server.packs.PackType
 import net.minecraft.server.packs.resources.PreparableReloadListener
@@ -29,4 +32,10 @@ fun registerClientResourceReloadListener(id: ResourceLocation, listener: Prepara
             override fun getFabricId() = id
         }
     )
+}
+
+fun sendHexicalKeyEvent(client: Minecraft, keyCode: Int, scanCode: Int, isPressed: Boolean) {
+    if (FabricLoader.getInstance().isModLoaded("hexical")) {
+        FabricHexDebugHexicalInterop.sendKeyEvent(client, keyCode, scanCode, isPressed)
+    }
 }